Replace TINYINT -> SMALLINT to avoid out of range error

+++ pkg/job_offer-9A645583F8F24B28/upgrade.sql
@@ -12,7 +12,7 @@ CREATE OR REPLACE TABLE job_offer (
     convocatoria VARCHAR(250),
     entidad_codigo SMALLINT,
     cierre DATE NOT NULL,
-    vacantes TINYINT,
+    vacantes SMALLINT,
     estudio TEXT,
     dependencia VARCHAR(10000),
     municipio VARCHAR(10000),
